[btree] Differentiate non-btree nodes from trashed blocks
It is useful if we could point out whether a node is trashed or not if
the checksum doesn't match, so users could easily know the severity of
errors. e.g., whether it is multiple activation, or there are damaged
disks.
