+++ src/pdata/btree.rs
@@ -262,4 +262,28 @@ pub fn calc_max_entries<V: Unpack>() -> usize {
     total / 3 * 3
 }
 
+// Verify the checksum of a node
+fn verify_checksum(b: &Block) -> std::result::Result<(), NodeError> {
+    use crate::checksum;
+    match checksum::metadata_block_type(b.get_data()) {
+        checksum::BT::NODE => Ok(()),
+        checksum::BT::UNKNOWN => Err(NodeError::ChecksumError),
+        _ => Err(NodeError::NotANode),
+    }
+}
+
+// Unpack a node and verify the checksum.
+// The returned error is content based. Context information is up to the caller.
+pub fn check_and_unpack_node<V: Unpack>(
+    b: &Block,
+    ignore_non_fatal: bool,
+    is_root: bool,
+) -> std::result::Result<Node<V>, NodeError> {
+    verify_checksum(b)?;
+    let node = unpack_node_raw::<V>(b.get_data(), ignore_non_fatal, is_root)?;
+    if node.get_header().block != b.loc {
+        return Err(NodeError::BlockNrMismatch);
+    }
+    Ok(node)
+}
 //------------------------------------------
